holy shit alot of things

+++ Code/main.py
@@ -1,8 +1,8 @@
 from PyQt6 import QtWidgets
 from PyQt6.QtWidgets import *
-from PyQt6.QtCore import Qt
+from PyQt6.QtCore import Qt, QDate
 from PyQt6 import uic
-from PyQt6.QtGui import QKeyEvent, QStandardItemModel, QStandardItem
+from PyQt6.QtGui import QKeyEvent, QStandardItemModel, QStandardItem, QTextCharFormat, QColor
 import os
 
 import time
@@ -11,6 +11,8 @@ from math import pi, sqrt
 
 
 app = QApplication(sys.argv)
+
+
 
 class MainWindow(QMainWindow):
     def __init__(self, name):
@@ -27,8 +29,12 @@ class MainWindow(QMainWindow):
         self.FileList.itemClicked.connect(self.clickedFileItem)
         self.Save.clicked.connect(self.save)
         self.SaveButton1.clicked.connect(self.save)
-        
-        
+        self.SaveName = ""
+        self.DeleteButton.clicked.connect(self.Delete)
+        self.NewFile.clicked.connect(self.newFile)
+        self.RefreshButton.clicked.connect(self.populateFileBrowser)
+        self.OpenWorkspaceButton.clicked.connect(self.openWorkspace)
+
         self.populateFileBrowser()
         self.CalendarButton.clicked.connect(self.openCalendarPage)
         self.JournalButton.clicked.connect(self.openJournalPage)
@@ -38,18 +44,65 @@ class MainWindow(QMainWindow):
         self.mdeditor.textChanged.connect(self.markdownUpdate)
         self.mdeditor.textChanged.connect(self.tempsave)
 
-
         
+        self.date = self.calendar.selectedDate()
+        #to get date in tuple form is self.date.getDate() returns in (yyyy, mm, dd)
         self.show()
 
+    def clearEdit(self):
+        self.mdeditor.setText("")
 
     def clickedFileItem(self, clickedItem):
         self.openFile = clickedItem.text()+".md"
         self.mdeditor.setText(self.filecontents[self.openFile])
-        
+    
+    def openWorkspace(self):
+        directory_path = QFileDialog.getExistingDirectory(parent=self, caption="Select Directory", directory="")
+        if directory_path:
+            self.wsPath = directory_path
+            self.populateFileBrowser()
+            self.clearEdit()
+
+    def newFile(self):
+        self.clearEdit()
+        self.openFile = ""
+
+
+    def Delete(self):
+            text, ok = QInputDialog.getText(self, "Input Dialog", "Do you want to delete "+"'"+self.openFile+"'"+"?(y/n)")
+            if ok and text:
+                if text == "y":
+                    del self.filecontents[self.openFile]
+                    os.remove(self.wsPath+"/"+self.openFile)
+                    self.populateFileBrowser()
+                    self.clearEdit()
+                    self.openFile = ""
+
+
+            elif ok and not text:
+                print("User entered nothing.")
+            else:
+                print("User cancelled the input.")
+
     def save(self):
-         with open(self.wsPath+"/"+self.openFile, "w") as openFile:
-             openFile.write(self.filecontents[self.openFile])
+        if self.openFile != "":
+            with open(self.wsPath+"/"+self.openFile, "w") as openFile:
+                openFile.write(self.filecontents[self.openFile])
+        else:
+            text, ok = QInputDialog.getText(self, "Input Dialog", "Enter Note Name:")
+            if ok and text:
+                self.filecontents[text+".md"] = self.mdeditor.toPlainText()
+                self.openFile = text+".md"
+                
+                self.save()
+                self.populateFileBrowser()
+
+            elif ok and not text:
+                print("User entered nothing.")
+            else:
+                print("User cancelled the input.")
+
+        
     def tempsave(self):
         self.filecontents[self.openFile] = self.mdeditor.toPlainText()
 
@@ -76,18 +129,21 @@ class MainWindow(QMainWindow):
 
 
     def populateFileBrowser(self, path= "PrePlanning"):
+        self.FileList.clear()
+        self.files = []
         filesAndFolders = os.listdir(self.wsPath)
+        
         for item in filesAndFolders:
             if item[-3:] == ".md":
                 self.files.append(item)
-        self.FileList.clear()
+        
         for file in self.files:
             item = QListWidgetItem(file[:-3])
             self.FileList.addItem(item)
             with open(self.wsPath+"/"+file, "r") as openFile:
                 self.filecontents[file] = openFile.read()
-                
-        print(self.filecontents)
+        
+
     def markdownUpdate(self):
         self.mdView.setMarkdown(self.mdeditor.toPlainText())
